Exam update done

+++ src/main/java/com/example/examsite/repository/TeacherRepository.java
@@ -3,6 +3,6 @@ package com.example.examsite.repository;
 import com.example.examsite.models.TeacherModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TeacherRepository extends JpaRepository<TeacherModel,Integer> {
+public interface TeacherRepository extends JpaRepository<TeacherModel,Long> {
     TeacherModel findByEmail(String email);
 }
